Start removing gulp packages and replacing them with webpack ones
