+++ .eslintrc
@@ -1,0 +1,13 @@
+{
+    "parser": "babel-eslint",
+    "env": {
+        "browser": true,
+        "node": true
+    },
+    "extends": "airbnb",
+    "rules": {
+        "react/sort-comp": 0,
+        "react/jsx-indent": [1, 4],
+        "react/jsx-indent-props": [1, 4]
+    }
+}
