--- src/Util.js
@@ -1,13 +1,0 @@
-/**
- * Get the version of Internet Explorer in use, or undefined
- *
- * @return {?number} ieVersion - IE version as an integer, or undefined if not IE
- */
-function getIEVersion() {
-    var match = navigator.userAgent.match(/(?:MSIE |Trident\/.*; rv:)(\d+)/);
-    return match ? parseInt(match[1]) : undefined;
-}
-
-module.exports = {
-    getIEVersion: getIEVersion,
-};
