+++ webpack.config.js
@@ -1,0 +1,46 @@
+var path = require('path');
+var webpack = require('webpack');
+var autoprefixer = require('autoprefixer');
+
+module.exports = {
+    entry: [
+        'webpack-dev-server/client?http://localhost:5000',
+        'webpack/hot/only-dev-server',
+        './src/index'
+    ],
+    output: {
+        path: __dirname,
+        filename: 'bundle.js',
+        publicPath: '/static/'
+    },
+    resolve: {
+        extensions: ['', '.js']
+    },
+    devtool: 'eval-source-map',
+    plugins: [
+        new webpack.HotModuleReplacementPlugin(),
+        new webpack.NoErrorsPlugin(),
+    ],
+    postcss: [
+        autoprefixer,
+    ],
+    module: {
+        loaders: [
+            {
+                test: /\.jsx?$/,
+                loaders: ['react-hot', 'babel'],
+                include: path.join(__dirname, 'src')
+            },
+            {
+                test: /\.scss$/,
+                loaders: [
+                    'style-loader',
+                    'css-loader?modules&importLoaders=1&localIdentName=[name]__[local]___[hash:base64:5]',
+                    'postcss-loader',
+                    'sass-loader',
+                ],
+                include: path.join(__dirname, 'src')
+            },
+        ],
+    }
+};
