+++ webpack.config.production.js
@@ -1,0 +1,51 @@
+var path = require('path');
+var webpack = require('webpack');
+var autoprefixer = require('autoprefixer');
+
+module.exports = {
+    entry: './src/index',
+    output: {
+        path: path.join(__dirname, 'dist'),
+        filename: 'bundle.js',
+        publicPath: '/'
+    },
+    resolve: {
+        extensions: ['', '.js']
+    },
+    devtool: 'source-map',
+    plugins: [
+        new webpack.optimize.OccurenceOrderPlugin(),
+        new webpack.DefinePlugin({
+            'process.env': {
+                'NODE_ENV': JSON.stringify('production')
+            },
+        }),
+        new webpack.optimize.UglifyJsPlugin({
+            compress: {
+                warnings: false
+            },
+        }),
+    ],
+    postcss: [
+        autoprefixer,
+    ],
+    module: {
+        loaders: [
+            {
+                test: /\.jsx?$/,
+                loaders: ['babel'],
+                include: path.join(__dirname, 'src')
+            },
+            {
+                test: /\.scss$/,
+                loaders: [
+                    'style-loader',
+                    'css-loader?modules&importLoaders=1&localIdentName=[name]__[local]___[hash:base64:5]',
+                    'postcss-loader',
+                    'sass-loader',
+                ],
+                include: path.join(__dirname, 'src')
+            },
+        ]
+    }
+};
